Move controlnet code to comfy/controlnet.py
